backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
asyncpg==0.29.0
aiohttp==3.9.5
python-dotenv==1.0.1
pydantic==2.6.4